split up histogram index
