+++ src/traces/histogram/index.js
@@ -10,14 +10,12 @@
 'use strict';
 
 var Plotly = require('../../plotly');
-var isNumeric = require('fast-isnumeric');
 
 /** 
  * Histogram has its own calc function,
  * but uses Bars.plot to display
  * and Bars.setPositions for stacking and grouping
  */
-var histogram = module.exports = {};
 
 /**
  * histogram errorBarsOK is debatable, but it's put in for backward compat.
@@ -25,7 +23,6 @@ var histogram = module.exports = {};
  * constant and % work but they're not so meaningful. I guess it could be cool
  * to allow quadrature combination of errors in summed histograms...
  */
-Plotly.Plots.register(Plotly.Bars, 'histogram',
     ['cartesian', 'bar', 'histogram', 'oriented', 'errorBarsOK', 'showLegend'], {
     description: [
         'The sample data from which statistics are computed is set in `x`',
@@ -37,276 +34,6 @@ Plotly.Plots.register(Plotly.Bars, 'histogram',
     ].join(' ')
 });
 
-Plotly.Plots.register(Plotly.Heatmap, 'histogram2d',
-    ['cartesian', '2dMap', 'histogram'], {
-    hrName: 'histogram_2d',
-    description: [
-        'The sample data from which statistics are computed is set in `x`',
-        'and `y` (where `x` and `y` represent marginal distributions,',
-        'binning is set in `xbins` and `ybins` in this case)',
-        'or `z` (where `z` represent the 2D distribution and binning set,',
-        'binning is set by `x` and `y` in this case).',
-        'The resulting distribution is visualized as a heatmap.'
-    ].join(' ')
-});
-
-histogram.attributes = require('./attributes');
-
-histogram.supplyDefaults = function(traceIn, traceOut) {
-    function coerce(attr, dflt) {
-        return Plotly.Lib.coerce(traceIn, traceOut, histogram.attributes, attr, dflt);
-    }
-
-    var binDirections = ['x'],
-        hasAggregationData,
-        x = coerce('x'),
-        y = coerce('y');
-
-    if(Plotly.Plots.traceIs(traceOut, '2dMap')) {
-        // we could try to accept x0 and dx, etc...
-        // but that's a pretty weird use case.
-        // for now require both x and y explicitly specified.
-        if(!(x && x.length && y && y.length)) {
-            traceOut.visible = false;
-            return;
-        }
-
-        // if marker.color is an array, we can use it in aggregation instead of z
-        hasAggregationData = coerce('z') || coerce('marker.color');
-
-        binDirections = ['x','y'];
-    } else {
-        var orientation = coerce('orientation', (y && !x) ? 'h' : 'v'),
-            sample = traceOut[orientation==='v' ? 'x' : 'y'];
-
-        if(!(sample && sample.length)) {
-            traceOut.visible = false;
-            return;
-        }
-
-        if(orientation==='h') binDirections = ['y'];
-
-        hasAggregationData = traceOut[orientation==='h' ? 'x' : 'y'];
-    }
-
-    if(hasAggregationData) coerce('histfunc');
-    coerce('histnorm');
-
-    binDirections.forEach(function(binDirection){
-        // data being binned - note that even though it's a little weird,
-        // it's possible to have bins without data, if there's inferred data
-        var binstrt = Plotly.Lib.coerce(traceIn, traceOut, histogram.attributes, binDirection + 'bins.start'),
-            binend = Plotly.Lib.coerce(traceIn, traceOut, histogram.attributes, binDirection + 'bins.end'),
-            autobin = binstrt && binend ? 
-                coerce('autobin' + binDirection, false) : 
-                coerce('autobin' + binDirection);
-
-        if(autobin) coerce('nbins' + binDirection);
-        else coerce(binDirection + 'bins.size');
-    });
-};
-
-var binFunctions = {
-    count: function(n, i, size) {
-        size[n]++;
-        return 1;
-    },
-
-    sum: function(n, i, size, counterData) {
-        var v = counterData[i];
-        if(isNumeric(v)) {
-            v = Number(v);
-            size[n] += v;
-            return v;
-        }
-        return 0;
-    },
-
-    avg: function(n, i, size, counterData, counts) {
-        var v = counterData[i];
-        if(isNumeric(v)) {
-            v = Number(v);
-            size[n] += v;
-            counts[n]++;
-        }
-        return 0;
-    },
-
-    min: function(n, i, size, counterData) {
-        var v = counterData[i];
-        if(isNumeric(v)) {
-            v = Number(v);
-            if(!isNumeric(size[n])) {
-                size[n] = v;
-                return v;
-            }
-            else if(size[n]>v) {
-                size[n] = v;
-                return v-size[n];
-            }
-        }
-        return 0;
-    },
-
-    max: function(n, i, size, counterData) {
-        var v = counterData[i];
-        if(isNumeric(v)) {
-            v = Number(v);
-            if(!isNumeric(size[n])) {
-                size[n] = v;
-                return v;
-            }
-            else if(size[n]<v) {
-                size[n] = v;
-                return v-size[n];
-            }
-        }
-        return 0;
-    }
-};
-
-function doAvg(size, counts) {
-    var nMax = size.length,
-        total = 0;
-    for(var i=0; i<nMax; i++) {
-        if(counts[i]) {
-            size[i] /= counts[i];
-            total += size[i];
-        }
-        else size[i] = null;
-    }
-    return total;
-}
-
-var normFunctions = {
-    percent: function(size, total) {
-        var nMax = size.length,
-            norm = 100/total;
-        for(var n=0; n<nMax; n++) size[n] *= norm;
-    },
-    probability: function(size, total) {
-        var nMax = size.length;
-        for(var n=0; n<nMax; n++) size[n] /= total;
-    },
-    density: function(size, total, inc, yinc) {
-        var nMax = size.length;
-        yinc = yinc||1;
-        for(var n=0; n<nMax; n++) size[n] *= inc[n]*yinc;
-    },
-    'probability density': function(size, total, inc, yinc) {
-        var nMax = size.length;
-        if(yinc) total/=yinc;
-        for(var n=0; n<nMax; n++) size[n] *= inc[n]/total;
-    }
-};
-
-histogram.calc = function(gd, trace) {
-    // ignore as much processing as possible (and including in autorange) if bar is not visible
-    if(trace.visible !== true) return;
-
-    // depending on orientation, set position and size axes and data ranges
-    // note: this logic for choosing orientation is duplicated in graph_obj->setstyles
-    var pos = [],
-        size = [],
-        i,
-        pa = Plotly.Axes.getFromId(gd,
-            trace.orientation==='h' ? (trace.yaxis || 'y') : (trace.xaxis || 'x')),
-        maindata = trace.orientation==='h' ? 'y' : 'x',
-        counterdata = {x: 'y', y: 'x'}[maindata];
-
-    // prepare the raw data
-    var pos0 = pa.makeCalcdata(trace, maindata);
-    // calculate the bins
-    if((trace['autobin' + maindata]!==false) || !(maindata + 'bins' in trace)) {
-        trace[maindata + 'bins'] = Plotly.Axes.autoBin(pos0, pa, trace['nbins' + maindata]);
-
-        // copy bin info back to the source data.
-        trace._input[maindata + 'bins'] = trace[maindata + 'bins'];
-    }
-
-    var binspec = trace[maindata + 'bins'],
-        allbins = typeof binspec.size === 'string',
-        bins = allbins ? [] : binspec,
-        // make the empty bin array
-        i2,
-        binend,
-        n,
-        inc = [],
-        counts = [],
-        total = 0,
-        norm = trace.histnorm,
-        func = trace.histfunc,
-        densitynorm = norm.indexOf('density')!==-1,
-        extremefunc = func==='max' || func==='min',
-        sizeinit = extremefunc ? null : 0,
-        binfunc = binFunctions.count,
-        normfunc = normFunctions[norm],
-        doavg = false,
-        rawCounterData;
-
-    if(Array.isArray(trace[counterdata]) && func!=='count') {
-        rawCounterData = trace[counterdata];
-        doavg = func==='avg';
-        binfunc = binFunctions[func];
-    }
-
-    // create the bins (and any extra arrays needed)
-    // assume more than 5000 bins is an error, so we don't crash the browser
-    i = binspec.start;
-    // decrease end a little in case of rounding errors
-    binend = binspec.end +
-        (binspec.start - Plotly.Axes.tickIncrement(binspec.start, binspec.size)) / 1e6;
-    while(i<binend && pos.length<5000) {
-        i2 = Plotly.Axes.tickIncrement(i, binspec.size);
-        pos.push((i + i2) / 2);
-        size.push(sizeinit);
-        // nonuniform bins (like months) we need to search,
-        // rather than straight calculate the bin we're in
-        if(allbins) bins.push(i);
-        // nonuniform bins also need nonuniform normalization factors
-        if(densitynorm) inc.push(1 / (i2 - i));
-        if(doavg) counts.push(0);
-        i = i2;
-    }
-
-    var nMax = size.length;
-    // bin the data
-    for(i=0; i<pos0.length; i++) {
-        n = Plotly.Lib.findBin(pos0[i], bins);
-        if(n>=0 && n<nMax) total += binfunc(n, i, size, rawCounterData, counts);
-    }
-
-    // average and/or normalize the data, if needed
-    if(doavg) total = doAvg(size, counts);
-    if(normfunc) normfunc(size, total, inc);
-
-    var serieslen = Math.min(pos.length, size.length),
-        cd = [],
-        firstNonzero = 0,
-        lastNonzero = serieslen-1;
-    // look for empty bins at the ends to remove, so autoscale omits them
-    for(i=0; i<serieslen; i++) {
-        if(size[i]) {
-            firstNonzero = i;
-            break;
-        }
-    }
-    for(i=serieslen-1; i>firstNonzero; i--) {
-        if(size[i]) {
-            lastNonzero = i;
-            break;
-        }
-    }
-
-    // create the "calculated data" to plot
-    for(i=firstNonzero; i<=lastNonzero; i++) {
-        if((isNumeric(pos[i]) && isNumeric(size[i]))) {
-            cd.push({p: pos[i], s: size[i], b: 0});
-        }
-    }
-
-    return cd;
-};
 
 histogram.calc2d = function(gd, trace) {
     var xa = Plotly.Axes.getFromId(gd, trace.xaxis||'x'),
